Rename GitHubUserDto to GitHubRepoDto and update properties to match repository data structure
